pyproject.toml
[project]
name = "data-pipelines"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy==1.26.0",
    "pandas==2.2.0",
    "polars>=1.29.0",
    "prefect==3.4.1",
    "prefect-docker>=0.6.4",
    "prefect-dbt==0.6.7",
    "dbt-core==1.8.1",
    "dbt-duckdb>=1.7.0",
    "dbt-clickhouse==1.8.0",
    "requests[socks]==2.31.0",
    "db-dtypes==1.1.1",
    "google-cloud-bigquery==3.14.1",
    "google-cloud-storage==2.14.0",
    "clickhouse-connect==0.8.16",
    "pydrive==1.3.1",
    "dbt-bigquery==1.8.0",
    "pydantic==2.10.1",
    "gspread>=6.2.1",
]
